Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,6 +5,6 @@ void main()
     int y;
     int z;
     scanf("%d%d%d",&x,&y,&z);
-    int avg= (x+y+z)/3;
-    pritnf("Average: %d",avg);
+    float avg= (x+y+z)/3;
+    pritnf("Average: %.2f",avg);
 }
